+++ lib/WorseReflection/Bridge/TolerantParser/Diagnostics/DocblockMissingReturnTypeProvider.php
@@ -19,7 +19,7 @@ use Phpactor\WorseReflection\Core\Util\NodeUtil;
  * Report when a method has a return type should be
  * augmented by a phpdoc.
  */
-class MissingDocblockReturnTypeProvider implements DiagnosticProvider
+class DocblockMissingReturnTypeProvider implements DiagnosticProvider
 {
     public function exit(NodeContextResolver $resolver, Frame $frame, Node $node): iterable
     {
@@ -77,7 +77,7 @@ class MissingDocblockReturnTypeProvider implements DiagnosticProvider
         }
 
         if ($actualReturnType->isClosure()) {
-            yield new MissingDocblockReturnTypeDiagnostic(
+            yield new DocblockMissingReturnTypeDiagnostic(
                 $method->nameRange(),
                 sprintf(
                     'Method "%s" is missing docblock return type: %s',
@@ -106,7 +106,7 @@ class MissingDocblockReturnTypeProvider implements DiagnosticProvider
             return;
         }
 
-        yield new MissingDocblockReturnTypeDiagnostic(
+        yield new DocblockMissingReturnTypeDiagnostic(
             $method->nameRange(),
             sprintf(
                 'Method "%s" is missing docblock return type: %s',
